@@ -42,27 +168,32 @@
   "homepage": "https://github.com/schoero/SwissQRBill#readme",
   "devDependencies": {
     "@types/blob-stream": "^0.1.30",
-    "@types/node": "^16.7.2",
+    "@types/node": "^16.10.1",
     "@types/qrcode": "^1.4.1",
     "@types/qrcode-svg": "^1.1.1",
     "@types/svg-parser": "^2.0.3",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
+    "@typescript-eslint/eslint-plugin": "^4.31.2",
+    "@typescript-eslint/parser": "^4.31.2",
+    "assert": "^2.0.0",
+    "brfs": "^2.0.2",
+    "browserify-zlib": "^0.2.0",
     "buffer": "^6.0.3",
     "eslint": "^7.32.0",
     "process": "^0.11.10",
-    "stream-browserify": "^3.0.0",
-    "ts-loader": "^9.2.5",
-    "typescript": "^4.3.5",
+    "readable-stream": "^3.6.0",
+    "transform-loader": "^0.2.4",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.4.3",
     "util": "^0.12.4",
-    "webpack": "^5.51.1",
+    "webpack": "^5.54.0",
     "webpack-cli": "^4.8.0"
   },
   "dependencies": {
     "@schoero/qrcode": "^1.4.5",
+    "@types/pdfkit": "^0.11.2",
     "blob-stream": "^0.1.3",
     "pdfkit": "^0.12.3",
-    "svg-parser": "^2.0.4",
+    "svg-engine": "^0.2.0",
     "svgpath": "^2.3.1"
   }
 }
