@@ -1,8 +1,9 @@
-const SwissQRBill = require("swissqrbill");
+import { PDF } from "swissqrbill/pdf";
 
 const data = {
   currency: "CHF",
   amount: 1199.95,
+  reference: "RF18539007547034",
   creditor: {
     name: "Robert Schneider AG",
     address: "Rue du Lac 1268",
@@ -20,4 +21,4 @@ const data = {
   }
 };
 
-const pdf = new SwissQRBill.PDF(data, "./output/normal-iban-no-reference.pdf");
+const pdf = new PDF(data, "./output/normal-iban-creditor-reference.pdf");
