@@ -1,15 +1,14 @@
-const SwissQRBill = require("swissqrbill");
+import { PDF } from "swissqrbill/pdf";
 
 const data = {
   currency: "CHF",
   amount: 1199.95,
-  reference: "210000000003139471430009017",
   creditor: {
     name: "Robert Schneider AG",
     address: "Rue du Lac 1268",
     zip: 2501,
     city: "Biel",
-    account: "CH4431999123000889012",
+    account: "CH5800791123000889012",
     country: "CH"
   },
   debtor: {
@@ -21,8 +20,4 @@ const data = {
   }
 };
 
-const pdf = new SwissQRBill.PDF(data, "./output/event.pdf");
-
-pdf.on("finish", () => {
-  console.log("File has been successfully created.");
-});
+const pdf = new PDF(data, "./output/normal-iban-no-reference.pdf");
