@@ -1,4 +1,5 @@
 const SwissQRBill = require("../");
+const { writeFileSync } = require("fs");
 
 const data = {
   currency: "CHF",
@@ -23,4 +24,6 @@ const data = {
   }
 };
 
-const pdf = new SwissQRBill.PDF(data, "./output/message-additional-information.pdf", { "size": "A6/5" });
+const pdf = new SwissQRBill.PDF(data, "./output/pdf/message-additional-information.pdf", { "size": "A6/5" });
+const svg = new SwissQRBill.SVG(data);
+writeFileSync("./output/svg/message-additional-information.svg", svg.toString());
