@@ -1,4 +1,5 @@
 const SwissQRBill = require("../");
+const { writeFileSync } = require("fs");
 
 const data = {
   currency: "CHF",
@@ -22,4 +23,6 @@ const data = {
   }
 };
 
-const pdf = new SwissQRBill.PDF(data, "./output/message.pdf", { "size": "A6/5" });
+const pdf = new SwissQRBill.PDF(data, "./output/pdf/message.pdf", { "size": "A6/5" });
+const svg = new SwissQRBill.SVG(data);
+writeFileSync("./output/svg/message.svg", svg.toString());
