@@ -1,4 +1,5 @@
 const SwissQRBill = require("../");
+const { writeFileSync } = require("fs");
 
 const data = {
   currency: "CHF",
@@ -13,4 +14,6 @@ const data = {
   }
 };
 
-const pdf = new SwissQRBill.PDF(data, "./output/no-debtor-no-amount.pdf");
+const pdf = new SwissQRBill.PDF(data, "./output/pdf/no-debtor-no-amount.pdf");
+const svg = new SwissQRBill.SVG(data);
+writeFileSync("./output/svg/no-debtor-no-amount.svg", svg.toString());
